+++ src/libraries/IntVeto/IntVetofa250Converter_factory.cc
@@ -24,7 +24,7 @@ jerror_t IntVetofa250Converter_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t IntVetofa250Converter_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
+jerror_t IntVetofa250Converter_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber)
 {
 	if (m_isFirstCallToBrun){
 			m_isFirstCallToBrun=0;
@@ -39,7 +39,7 @@ jerror_t IntVetofa250Converter_factory::brun(jana::JEventLoop *eventLoop, int ru
 //------------------
 // evnt
 //------------------
-jerror_t IntVetofa250Converter_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t IntVetofa250Converter_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 
 	// Code to generate factory data goes here. Add it like:
